Break up sql module into submodules
